[Adjustment][Jhoanna] BuyingItemList
